Remove DEFAULT_FILENAME consts
Usage was not consistent.
Filenames are never optional in component constructors.
By default, when no filename is given for a read or write function,
it means that the constructor value is used.

Only vector.read works differently, passing None means don't use it for reading.
vector doesn't ask for a filename in the constructor.