Use fetch-depth 0 in release prep flow
During my testing no tags were detected, which I thought was because there were no tags on the fork I was using. When I was working on the `v1.1.0` release I noticed that this was also true in the main repo. I think now it's because the history doesn't get fetched by a checkout by default. Using fetch-depth 0 pulls in the full history, so the commands (that work locally) should function as intended with after this