Safely close all xr.DataSet objects during write
When writing, we need to ensure that all netcdf DataSets are closed.
Otherwise we are not able to overwrite the files.

read_nc and read_ncs now always return DataSet.
It's upon the caller to maintain its reference for closing,
and to split it into a DataArray if required.

Fixes #778