Close xarray before finishing DeferredFileClose (#1254)
Fixes (#778)

Xarray keeps Datasets open and does not support concurrent reading and writing to the same file. This gives conflicts in Model.update(). Therefore, all open Datasets need to be tracked within each component. Writing is first done to a temporary file next to the target. Reading can proceed as normal. After that, all open datasets are closed with ModelComponent.cleanup(). And the model can finish writing by calling ModelComponent.finish_write() right after to clean up all DeferredFileClose objects.

This solves the problem with open xarray Datasets. Bookkeeping of these Datasets is hard and also requires the caller of the code to be careful with keeping datasets open. If the user calls python functions like ModelComponent.set, hydromt doesn't know if it is a netcdf file underneath and if it should close it. Therefore, the end user is responsible in that case and we will log a PermissionError to the user during build or update.

This functionality does not yet change the way that the DataCatalog opens xarray objects and the bookkeeping there is not yet aligned. There will be similar cases where the original Dataset should actually be closed.

It is important to notice that retrieving a DataArray from a Dataset does not copy over the close handler (Dataset._close). Therefore, the original Dataset always needs to be in the bookkeeping somewhere. For the components it is possible to do this by adding them to ModelComponent._open_datasets.